requirements.txt
transformers
torch
huggingface_hub[hf_xet]
protobuf
google-generativeai
groq
python-dotenv
streamlit
beautifulsoup4
sentence-transformers
pinecone
langchain
pinecone[grpc]
plotly
